converter_service/requirements.txt
Django>=4.1,<5.0
psycopg2>=2.8
djangorestframework
markdown
django-filter
boto3
botocore
